+++ src/Ads/Widgets/WidgetInterface.php
@@ -4,5 +4,5 @@ namespace BackendApp\Ads\Widgets;
 
 interface WidgetInterface
 {
-	public function getPointsValue(array $widget) : int;
+	public function getPointsValue(array $widget) : float;
 }
